requirements.txt
matplotlib==3.5.3
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
igraph==0.10.6
PyYAML==6.0
tqdm
scikit-learn
